Refactoring the color representation
* Move the existing code to a new color module
* Add doctests
* Add `Display` trait and `PartialEq` trait implementations
