+++ src/color/mod.rs
@@ -1,0 +1,4 @@
+#[allow(clippy::module_inception)]
+mod color;
+
+pub use color::Color;
